Add file output support and a default input order.
The program will use the courses and preferences files if provided in cli, otherwise it will look for `json` and `csv` files in the input folder (default `input`). `csv` files are preferred over `json`. `generate_random_presets.py` will now generate the presets in the `input` folder.

Add `format` folder with examples for the program's formats.